# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finanza Personale API",
    "description": "API REST per leggere profilo, strategia, snapshot del portafoglio e transazioni di un tracker di finanza personale. Pensata per agenti AI: chiama prima /profile per il contesto, poi /summary per i saldi e le posizioni, infine /transactions per i dettagli.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://finanza-personale-cris.vercel.app/api/v1",
      "description": "Produzione"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/transactions": {
      "get": {
        "operationId": "getTransactions",
        "summary": "Lista transazioni paginata",
        "description": "Lista transazioni ordinata dalla più recente. Filtrabile per tipo, per conto e per intervallo di date (from/to, usabili anche singolarmente).",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Risultati per pagina (max 1000).",
            "required": false,
            "schema": { "type": "integer", "default": 100, "maximum": 1000 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset per la paginazione.",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filtra per tipo di transazione.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "BUY",
                "SELL",
                "DEPOSIT",
                "WITHDRAW",
                "EXPENSE",
                "INCOME",
                "TRANSFER_IN",
                "TRANSFER_OUT",
                "FEE",
                "DIVIDEND"
              ]
            }
          },
          {
            "name": "account",
            "in": "query",
            "description": "Filtra per nome conto (es. Trade Republic, BCC, Revolut, Binance).",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "account_id",
            "in": "query",
            "description": "Filtra per id (UUID) del conto. Alternativa più precisa a 'account' per nome.",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Data minima inclusa (formato YYYY-MM-DD). Restituisce le transazioni con date >= from. Può essere usata da sola o insieme a 'to'.",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Data massima inclusa (formato YYYY-MM-DD). Restituisce le transazioni con date <= to. Può essere usata da sola o insieme a 'from'.",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista transazioni",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Transaction" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token API passato come header Authorization: Bearer <token>."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Token mancante o non valido",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "ServerError": {
        "description": "Errore del server",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "date": { "type": "string", "format": "date" },
          "type": { "type": "string" },
          "quantity": { "type": "number", "nullable": true },
          "amount_eur": { "type": "number" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "accounts": {
            "type": "object",
            "nullable": true,
            "properties": {
              "name": { "type": "string" }
            }
          },
          "assets": {
            "type": "object",
            "nullable": true,
            "properties": {
              "ticker": { "type": "string" },
              "type": { "type": "string" }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "status": { "type": "integer" }
        }
      }
    }
  }
}